resolve SPDX URL & refactoring.
